Split columns_ui_appearance.h into multiple files
