+++ colours.cpp
@@ -1,0 +1,29 @@
+#include "ui_extension.h"
+
+namespace cui::colours {
+
+#ifdef _WIN64
+
+const GUID manager::class_guid = {0x7e060bd9, 0xb835, 0x4a4c, {0xaa, 0xc5, 0xdd, 0xdb, 0x9e, 0x23, 0x86, 0x12}};
+
+const GUID manager_instance::class_guid
+    = {0xf0fe4280, 0xf056, 0x426f, {0x89, 0x81, 0x22, 0x15, 0xe4, 0x10, 0x97, 0x45}};
+
+const GUID client::class_guid = {0x8eb06771, 0x9e10, 0x4636, {0xa4, 0x85, 0x14, 0x74, 0x78, 0x12, 0x2f, 0x7c}};
+
+#else
+
+const GUID manager::class_guid = {0xec705736, 0x4c68, 0x44fe, {0xb2, 0x40, 0x74, 0x79, 0x3b, 0x40, 0xff, 0xc0}};
+
+const GUID manager_instance::class_guid = {0xe1979e50, 0xb496, 0x47e9, {0x88, 0xcc, 0x4, 0xff, 0xf1, 0x90, 0x24, 0xd0}};
+
+const GUID client::class_guid = {0xa4ae2971, 0x64c8, 0x4118, {0x8e, 0x2a, 0xef, 0xdc, 0x37, 0xa5, 0x93, 0x72}};
+
+#endif
+
+bool is_dark_mode_active()
+{
+    return helper().is_dark_mode_active();
+}
+
+} // namespace cui::colours
