+++ columns_ui-sdk.vcxproj
@@ -163,6 +163,8 @@
   <ItemGroup>
     <ClInclude Include="base.h" />
     <ClInclude Include="buttons.h" />
+    <ClInclude Include="colours.h" />
+    <ClInclude Include="fonts.h" />
     <ClInclude Include="font_manager_v3.h" />
     <ClInclude Include="menu.h" />
     <ClInclude Include="splitter.h" />
@@ -172,12 +174,14 @@
     <ClInclude Include="win32_helpers.h" />
     <ClInclude Include="window_helper.h" />
     <ClInclude Include="columns_ui.h" />
-    <ClInclude Include="columns_ui_appearance.h" />
     <ClInclude Include="ui_extension.h" />
     <ClInclude Include="container_uie_window_v3.h" />
     <ClInclude Include="container_window_v3.h" />
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="colours.cpp" />
+    <ClCompile Include="fonts.cpp" />
+    <ClCompile Include="font_manager_v3.cpp" />
     <ClCompile Include="win32_helpers.cpp" />
     <ClCompile Include="window_helper.cpp" />
     <ClCompile Include="columns_ui.cpp" />
