+++ columns_ui.cpp
@@ -16,21 +16,6 @@ const GUID group::class_guid = {0xdf670755, 0x3cac, 0x4d3a, {0x88, 0x12, 0xea, 0
 
 const GUID control::class_guid = {0x8e0548d4, 0xf801, 0x440c, {0x81, 0xb1, 0xc2, 0x74, 0x30, 0x73, 0x08, 0x4a}};
 
-const GUID colours::manager::class_guid
-    = {0x7e060bd9, 0xb835, 0x4a4c, {0xaa, 0xc5, 0xdd, 0xdb, 0x9e, 0x23, 0x86, 0x12}};
-
-const GUID colours::manager_instance::class_guid
-    = {0xf0fe4280, 0xf056, 0x426f, {0x89, 0x81, 0x22, 0x15, 0xe4, 0x10, 0x97, 0x45}};
-
-const GUID fonts::manager::class_guid = {0x3e5a965d, 0xe161, 0x45b1, {0x98, 0xf9, 0xcf, 0xa1, 0x49, 0xd9, 0xea, 0x43}};
-
-const GUID fonts::manager_v2::class_guid
-    = {0x82534e1a, 0x1c15, 0x4386, {0x8d, 0xe2, 0x11, 0x00, 0x98, 0x3b, 0x7a, 0xd3}};
-
-const GUID colours::client::class_guid = {0x8eb06771, 0x9e10, 0x4636, {0xa4, 0x85, 0x14, 0x74, 0x78, 0x12, 0x2f, 0x7c}};
-
-const GUID fonts::client::class_guid = {0x6632d984, 0xc1d3, 0x477b, {0x8e, 0x14, 0x38, 0xc8, 0xa0, 0x25, 0xea, 0xde}};
-
 #else
 
 namespace fcl {
@@ -45,31 +30,7 @@ const GUID group::class_guid = {0x78deea88, 0xa3cb, 0x47c7, {0x95, 0xde, 0x39, 0
 
 const GUID control::class_guid = {0x28992df0, 0x1ee2, 0x4e06, {0xac, 0x86, 0xbf, 0x4f, 0x26, 0x6e, 0x2e, 0x56}};
 
-const GUID colours::manager::class_guid
-    = {0xec705736, 0x4c68, 0x44fe, {0xb2, 0x40, 0x74, 0x79, 0x3b, 0x40, 0xff, 0xc0}};
-
-const GUID colours::manager_instance::class_guid
-    = {0xe1979e50, 0xb496, 0x47e9, {0x88, 0xcc, 0x4, 0xff, 0xf1, 0x90, 0x24, 0xd0}};
-
-const GUID fonts::manager::class_guid = {0x5b277e9, 0x83c3, 0x4f75, {0xb9, 0x2f, 0x66, 0xc1, 0x4, 0x9d, 0xb, 0x68}};
-
-const GUID fonts::manager_v2::class_guid
-    = {0x6dcfd77f, 0x891, 0x4fc0, {0x85, 0x74, 0x4e, 0xe3, 0x91, 0x9b, 0xd4, 0x6a}};
-
-const GUID colours::client::class_guid = {0xa4ae2971, 0x64c8, 0x4118, {0x8e, 0x2a, 0xef, 0xdc, 0x37, 0xa5, 0x93, 0x72}};
-
-const GUID fonts::client::class_guid = {0x3fbcc2b0, 0x978e, 0x406f, {0xa4, 0x46, 0x46, 0x77, 0xe0, 0xd4, 0xc5, 0x8e}};
-
 #endif
-
-namespace fonts {
-
-#ifndef CUI_SDK_DWRITE_DISABLED
-const GUID font::class_guid{0x907a8df2, 0xc398, 0x402b, {0x83, 0x1f, 0x03, 0x11, 0xf5, 0xa9, 0x8f, 0x19}};
-const GUID manager_v3::class_guid{0xdb18291d, 0x8013, 0x4f32, {0x96, 0x4b, 0x37, 0x77, 0xfd, 0xba, 0x56, 0x90}};
-#endif
-
-} // namespace fonts
 
 void cui::fcl::dataset::get_data_to_array(pfc::array_t<uint8_t>& p_data, t_uint32 type, t_export_feedback& feedback,
     abort_callback& p_abort, bool b_reset) const
@@ -85,21 +46,4 @@ void cui::fcl::dataset::set_data_from_ptr(
     return set_data(&reader, size, type, feedback, p_abort);
 }
 
-bool colours::is_dark_mode_active()
-{
-    return colours::helper().is_dark_mode_active();
-}
-
-bool fonts::client::create_by_guid(const GUID& p_guid, client::ptr& p_out)
-{
-    service_enum_t<client> p_enum;
-    client::ptr ptr;
-    while (p_enum.next(ptr)) {
-        if (ptr->get_client_guid() == p_guid) {
-            p_out = ptr;
-            return true;
-        }
-    }
-    return false;
-}
 } // namespace cui
