+++ font_manager_v3.cpp
@@ -1,0 +1,10 @@
+#include "ui_extension.h"
+
+namespace cui::fonts {
+
+#ifndef CUI_SDK_DWRITE_DISABLED
+const GUID font::class_guid{0x907a8df2, 0xc398, 0x402b, {0x83, 0x1f, 0x03, 0x11, 0xf5, 0xa9, 0x8f, 0x19}};
+const GUID manager_v3::class_guid{0xdb18291d, 0x8013, 0x4f32, {0x96, 0x4b, 0x37, 0x77, 0xfd, 0xba, 0x56, 0x90}};
+#endif
+
+} // namespace cui::fonts
